Internal functions and functionality

Types

QEDFeynmanDiagrams.AbstractTreeLevelFeynmanDiagram — Type

AbstractTreeLevelFeynmanDiagram

Abstract base type for FeynmanDiagrams. Must implement the functions

process(::AbstractTreeLevelFeynmanDiagram)::QEDbase.AbstractProcessDefinition
virtual_particles(::AbstractTreeLevelFeynmanDiagram)::NTuple{N, Tuple{QEDbase.AbstractParticleType, BitArray}}

By using the QEDbase.AbstractProcessDefinition interface, the function external_particles is automatically provided.

For more information on what the interface functions should do, see their documentation: process, virtual_particles

Functions

QEDFeynmanDiagrams.disjunct — Function

disjunct(a::VirtualParticle, b::VirtualParticle)

Return true if the momenta contributions of a and b are disjunct.

QEDFeynmanDiagrams.contains — Function

contains(a::VirtualParticle, b::VirtualParticle)

Returns true if the set of particles contributing to a are contains the set of particles contributing to b.

QEDFeynmanDiagrams.make_up — Function

make_up(a::VirtualParticle, b::VirtualParticle, c::VirtualParticle)

For virtual particles a, b, and c, return true if a and b's joint momentum contributions add up to c's momentum contributions.

QEDFeynmanDiagrams.are_total — Function

are_total(a::VirtualParticle, b::VirtualParticle, c::VirtualParticle)

Return true if a, b and c combined contain all external particles exactly once.

QEDFeynmanDiagrams._pseudo_virtual_particles — Function

_pseudo_virtual_particles

Return a vector of VirtualParticle for each external particle. These are not actually virtual particles, but can be helpful as entry points.